Adds script to set MacTahoe Light theme
Adds a script to configure the system theme to MacTahoe Light.

This involves disabling extensions associated with the Fluent theme and enabling those used by MacTahoe.
It configures various dconf settings related to interface, GTK, icon themes,
accent colors and cursor themes to achieve the desired look.

The script also sets configurations for the accent-gtk-theme, accent-user-theme and accent-icons-theme extensions,
including light and dark variations for different color schemes.

Finally, it configures the light-dark-cursor-theme extension and sets the panel position for the openweatherrefined extension.