Adds script to set MacTahoe Solid Light theme
Adds a script that configures the system to use the MacTahoe Solid Light theme.

This script disables extensions conflicting with the theme, enables
theme-specific extensions, and sets various dconf settings to
apply the theme and related customizations for GTK, icon, cursor themes,
and shell extensions related to accent colors and theme settings. It also
configures the position of the OpenWeatherRefined extension, and sets the
appearance of the space-bar extension.

It also includes a commented-out section to restart gnome-shell (for X11)
and a gnome-session-quit command (for Wayland), to apply changes requiring
a restart.