Adds script to set WhiteSur Solid Light theme
Creates a script to configure the desktop environment for the WhiteSur Solid Light theme.

This script disables extensions associated with the Fluent theme and enables those used by WhiteSur.

It also sets various dconf settings to configure the interface, including:

- Theme, icon theme, and cursor theme
- Settings for accent-gtk-theme, accent-user-theme, and accent-icons-theme
- Light and dark cursor theme
- Position in panel for openweatherrefined
- Space-bar appearance

The script also attempts to restart gnome-shell (X11 only) or prompts to logout.