Improves installer script functionality and logging
Adds support for btrfs with LVM during installation, allowing for logical volume management with btrfs filesystems.

Updates logging to append to the log file instead of overwriting, preserving the install history. Also, it generates the log file name with date and time to avoid name collisions.

Modifies grub configuration to update the logo, distributor name, and enable os prober in the grub menu.